refactor(bot): move handlers to commands subdirectory
Reorganize handlers into bot/handlers/commands/ to match autochecker
pattern expectation (bot/handlers/**/*.py).

Update imports in bot.py to use handlers.commands.*
