+++ bot/handlers/commands/help.py
